requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0

# Collecte
feedparser>=6.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0
lxml>=5.0.0

# IA / Analyse
anthropic>=0.40.0
tiktoken>=0.5.0

# Email
jinja2>=3.1.0

# Stockage
cryptography>=41.0.0

# Scheduling (optionnel, pour exécution standalone)
apscheduler>=3.10.0
